CDDSO-203: uncommitted changes on disk

+++ cdds/cdds/convert/concatenation/concatenation_setup.py
@@ -55,7 +55,7 @@ def organise_concatenations(reference_date, start_date, end_date,
         Relates each output file to the list of files that should be
         concatenated to create it.
     """
-    reinitialisation_days = Calendar.default().DAYS_IN_YEAR * reinitialisation_years
+    reinitialisation_days = int(Calendar.default().DAYS_IN_YEAR * reinitialisation_years)
     reinit_duration = Duration(days=reinitialisation_days)
     chunk_start = reference_date
     chunk_end = chunk_start + reinit_duration
